requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
cfgv==3.4.0
distlib==0.3.8
fastapi==0.110.0
filelock==3.13.1
greenlet==3.0.3
identify==2.5.35
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.6.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.3
sentry-sdk==1.41.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
